+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- MSFT_xIPAddressOption:
+  - Added a new resource to set the SkipAsSource option for an IP address.
 - MSFT_xWeakHostSend:
   - Created the new Weak Host Send resource.
 - MSFT_xWeakHostReceive: